+++ Contents/mods/ChaosMod/42/media/lua/client/ChaosMod/Effects/List/NPC/EffectSpawnGrieferSanta.lua
@@ -32,6 +32,8 @@ function EffectSpawnGrieferSanta:OnStart()
 
     npc.npcGroup = ChaosNPCGroupID.RAIDERS
 
-    npc:SetWeapon("Base.Sledgehammer")
+    -- npc:SetWeapon("Base.Sledgehammer")
     npc:EnterPlayerVehicle(player)
+
+    npc:SetWeapon("Base.AssaultRifle")
 end